pyproject.toml
[tool.poetry]
name = "wedding-bot"
version = "0.1.0"
description = "My telegram wedding bot"
authors = ["GrigoriyKruchinin <GrigoriyKruchinin>"]
readme = "README.md"
packages = [{include = "bot"}]

[tool.poetry.dependencies]
python = "^3.8.1"
aiogram = "^3.3.0"
python-dotenv = "^1.0.0"
httpx = "^0.26.0"
translate = "^3.6.1"
pydantic = "^2.5.3"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
